RE-CREATE Stripe COMPONENT
- in client/src/components/Stripe/
